Latest package for constant URL
This package is the latest version of 1.0.1.6, which is the same as the
one named WechatHistory_1.0.1.6.zip. Uploading this is to maintain a
constant URL for downloading, which is used in the article in my blog.
From now on, I don't need to update the latest URL in the article
whenever I upload a new version.