Add try-catch in GetAllPYLetters and output the exception message
None of exception caused by this function is handled in former edition. Lots of user have different type of names and charset, which often cause the exception. Thus the exception will be caputured and the exception message will be output as a messagebox. The user will be adviced to send the message to the author for better debugging.